+++ crates/common/src/config/mailstore/jmap.rs
@@ -23,10 +23,10 @@ pub struct JmapConfig {
     pub get_max_objects: usize,
     pub set_max_objects: usize,
 
-    pub upload_max_size: usize,
+    pub upload_max_size: u64,
     pub upload_max_concurrent: Option<u64>,
 
-    pub upload_tmp_quota_size: usize,
+    pub upload_tmp_quota_size: u64,
     pub upload_tmp_quota_amount: usize,
     pub upload_tmp_ttl: u64,
 
@@ -79,9 +79,9 @@ impl JmapConfig {
             request_max_concurrent: jmap.max_concurrent_requests,
             get_max_objects: jmap.get_max_results as usize,
             set_max_objects: jmap.set_max_objects as usize,
-            upload_max_size: jmap.max_upload_size as usize,
+            upload_max_size: jmap.max_upload_size,
             upload_max_concurrent: jmap.max_concurrent_uploads,
-            upload_tmp_quota_size: jmap.upload_quota as usize,
+            upload_tmp_quota_size: jmap.upload_quota,
             upload_tmp_quota_amount: jmap.max_upload_count as usize,
             upload_tmp_ttl: jmap.upload_ttl.into_inner().as_secs(),
             mail_parse_max_items: jmap.parse_limit_email as usize,
